# src/generated/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fallback",
    "version": "0.1.0-beta.1",
    "description": "Small pay-per-call recovery utilities for autonomous agents. Don't spend a dollar of reasoning on a cent-sized problem.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://fallback.factrail.online"
    }
  ],
  "paths": {
    "/v1/tools/source_route": {
      "post": {
        "operationId": "source_route",
        "summary": "source_route",
        "description": "Find likely authoritative or machine-readable sources when you know what information you need but not where to retrieve it. Provide a goal and optional publisher domain or start URL; it returns ranked candidate routes and bounded evidence. Use it before blind searching. Do not use it when you already have the correct source or only need reasoning over supplied context.",
        "tags": [
          "source-discovery"
        ],
        "x-fallback-tool-id": "source_route",
        "x-fallback-status": "available",
        "x-fallback-price-usd": "0.02",
        "x-fallback-x402-active": true,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "goal": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  },
                  "domain": {
                    "type": "string",
                    "maxLength": 253,
                    "format": "hostname",
                    "pattern": "^(?=.{1,253}\\.?$)[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[-0-9a-zA-Z]{0,61}[0-9a-zA-Z])?)*\\.?$"
                  },
                  "start_url": {
                    "type": "string",
                    "maxLength": 2048
                  },
                  "preferred_formats": {
                    "maxItems": 7,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "json",
                        "csv",
                        "xml",
                        "rss",
                        "api",
                        "bulk_download",
                        "html"
                      ]
                    }
                  },
                  "require_official": {
                    "default": false,
                    "type": "boolean"
                  },
                  "max_candidates": {
                    "default": 5,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  }
                },
                "required": [
                  "goal"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool response envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment required for this available paid tool."
          },
          "404": {
            "description": "Tool is unknown or unavailable."
          }
        }
      }
    },
    "/v1/tools/error_route": {
      "post": {
        "operationId": "error_route",
        "summary": "error_route",
        "description": "Diagnose a failed API, HTTP, MCP, or tool request and return a structured, bounded next action. Use it after a request fails instead of repeatedly retrying or spending a large reasoning loop diagnosing common errors. Provide a status, response, or error text; it does not execute requests or verify undocumented fixes.",
        "tags": [
          "recovery"
        ],
        "x-fallback-tool-id": "error_route",
        "x-fallback-status": "available",
        "x-fallback-price-usd": "0.002",
        "x-fallback-x402-active": true,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "goal": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "request": {
                    "type": "object",
                    "properties": {
                      "method": {
                        "type": "string",
                        "maxLength": 16
                      },
                      "url": {
                        "type": "string",
                        "maxLength": 2048
                      },
                      "headers": {
                        "type": "object",
                        "propertyNames": {
                          "type": "string",
                          "maxLength": 128
                        },
                        "additionalProperties": {
                          "type": "string",
                          "maxLength": 2048
                        }
                      },
                      "body": {}
                    },
                    "additionalProperties": false
                  },
                  "response": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "minimum": 100,
                        "maximum": 599
                      },
                      "headers": {
                        "type": "object",
                        "propertyNames": {
                          "type": "string",
                          "maxLength": 128
                        },
                        "additionalProperties": {
                          "type": "string",
                          "maxLength": 2048
                        }
                      },
                      "body": {
                        "type": "string",
                        "maxLength": 8192
                      }
                    },
                    "additionalProperties": false
                  },
                  "error": {
                    "type": "string",
                    "maxLength": 2048
                  },
                  "attempt": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool response envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment required for this available paid tool."
          },
          "404": {
            "description": "Tool is unknown or unavailable."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ToolResponse": {
        "type": "object",
        "required": [
          "success",
          "toolId",
          "toolVersion",
          "requestId",
          "execution"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "toolId": {
            "type": "string"
          },
          "toolVersion": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "result": {},
          "error": {
            "type": "object"
          },
          "execution": {
            "type": "object"
          }
        }
      }
    }
  }
}